+++ CHANGELOG.md
@@ -14,7 +14,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - `demo1()`
 - envelopes module: `ADSR_envelope()` function
 - signals module: `add_karplus_strong()` for plucked strings
-- sound_effects module: `apply_delay_effect()`
+- sound_effects module: 
+  - `apply_delay_effect()`
+  - `apply_fuzz_effect()`
 
 ## [Forsynth 0.1 "Stockhausen"] 2021-02-19